:boom: Add chainage to enigme
