+++ src/pages/creer-une-enigme/1.tsx
@@ -1,13 +1,11 @@
 import { useRouter } from 'next/router';
 import React from 'react';
 
-import { Button, TextField } from '@material-ui/core';
-
 import { isEnigme } from 'src/activity-types/anyActivity';
-import { ENIGME_DATA, ENIGME_TYPES } from 'src/activity-types/enigme.const';
-import { EnigmeData } from 'src/activity-types/enigme.types';
+import { ENIGME_TYPES } from 'src/activity-types/enigme.const';
 import { Base } from 'src/components/Base';
 import { Steps } from 'src/components/Steps';
+import { ActivitySelect } from 'src/components/activities/ActivitySelect';
 import { BackButton } from 'src/components/buttons/BackButton';
 import { ThemeChoiceButton } from 'src/components/buttons/ThemeChoiceButton';
 import { ActivityContext } from 'src/contexts/activityContext';
@@ -17,106 +15,97 @@ import { ActivityType } from 'types/activity.type';
 const EnigmeStep1: React.FC = () => {
   const router = useRouter();
   const { activity, createNewActivity, updateActivity } = React.useContext(ActivityContext);
+  const selectRef = React.useRef<HTMLDivElement>(null);
+
+  // enigme sub-type
   const enigmeTypeIndex =
     activity !== null && 'edit' in router.query && isEnigme(activity)
       ? activity.subType ?? 0
       : parseInt(getQueryString(router.query['category']) ?? '-1', 10) ?? 0;
   const enigmeType = ENIGME_TYPES[enigmeTypeIndex] ?? ENIGME_TYPES[0];
-  const enigmeData = ENIGME_DATA[enigmeTypeIndex] ?? ENIGME_DATA[0];
 
-  const data = (activity?.data as EnigmeData) || null;
-
-  const onClick = (index: number) => () => {
-    if (index === -1) {
-      if (!data.themeName) {
-        return;
-      }
-      updateActivity({ data: { ...data, theme: index, themeName: data.themeName.toLowerCase() } });
-    } else {
-      const newData = data;
-      delete newData.themeName;
-      updateActivity({ data: { ...newData, theme: index } });
+  const onNext = (clear: boolean) => () => {
+    if (clear) {
+      updateActivity({ responseActivityId: null, responseType: null });
     }
     router.push('/creer-une-enigme/2');
+  };
+  const onChange = (id: number | null, type: ActivityType | null) => {
+    if (activity !== null) {
+      updateActivity({ responseActivityId: id, responseType: type });
+    }
   };
 
   const created = React.useRef(false);
   React.useEffect(() => {
     if (!created.current) {
+      const responseActivityId = 'responseActivityId' in router.query ? parseInt(getQueryString(router.query.responseActivityId), 10) ?? null : null;
+      const responseActivityType =
+        'responseActivityType' in router.query ? parseInt(getQueryString(router.query.responseActivityType), 10) ?? null : null;
       if (!('edit' in router.query)) {
         created.current = true;
-        createNewActivity(ActivityType.ENIGME, enigmeTypeIndex, {
-          theme: 0,
-          indiceContentIndex: 1,
-          timer: 0,
-        });
+        createNewActivity(
+          ActivityType.ENIGME,
+          enigmeTypeIndex,
+          {
+            theme: 0,
+            indiceContentIndex: 1,
+            timer: 0,
+          },
+          responseActivityId,
+          responseActivityType,
+        );
+        if (responseActivityId !== null) {
+          if (selectRef.current) {
+            selectRef.current.scrollIntoView({ behavior: 'smooth' });
+          }
+        }
       } else if (activity && (!isEnigme(activity) || activity.subType !== enigmeTypeIndex)) {
         created.current = true;
-        createNewActivity(ActivityType.ENIGME, enigmeTypeIndex, {
-          theme: 0,
-          indiceContentIndex: 1,
-          timer: 0,
-        });
+        createNewActivity(
+          ActivityType.ENIGME,
+          enigmeTypeIndex,
+          {
+            theme: 0,
+            indiceContentIndex: 1,
+            timer: 0,
+          },
+          responseActivityId,
+          responseActivityType,
+        );
+        if (responseActivityId !== null) {
+          if (selectRef.current) {
+            selectRef.current.scrollIntoView({ behavior: 'smooth' });
+          }
+        }
       }
     }
   }, [activity, createNewActivity, enigmeTypeIndex, router]);
+
+  if (!activity) {
+    return (
+      <Base>
+        <div></div>
+      </Base>
+    );
+  }
 
   return (
     <Base>
       <div style={{ width: '100%', padding: '0.5rem 1rem 1rem 1rem' }}>
         {!('edit' in router.query) && <BackButton href="/creer-une-enigme" />}
         <Steps
-          steps={['Choix de la catégorie', enigmeType.step2 ?? "Description de l'objet", "Création de l'indice", 'Prévisualisation']}
+          steps={['Démarrer', 'Choix de la catégorie', enigmeType.step2 ?? "Description de l'objet", "Création de l'indice", 'Prévisualisation']}
           activeStep={0}
         />
         <div className="width-900">
-          <h1>{enigmeType.titleStep1}</h1>
-          <p className="text" style={{ fontSize: '1.1rem' }}>
-            {enigmeType.description}
-          </p>
-          <div>
-            {enigmeData.map((t, index) => (
-              <ThemeChoiceButton key={index} label={t.label} description={t.description} onClick={onClick(index)} />
-            ))}
-            <ThemeChoiceButton
-              additionalContent={
-                <div className="text-center">
-                  <h3>Donnez un nom à la catégorie :</h3>
-                  <div style={{ display: 'inline-flex', alignItems: 'center', margin: '0.5rem 0' }}>
-                    <span style={{ marginRight: '0.3rem' }}>{`${enigmeType.title2} est`}</span>
-                    {data !== null && (
-                      <TextField
-                        value={data.themeName || ''}
-                        onChange={(event) => {
-                          updateActivity({ data: { ...data, themeName: event.target.value } });
-                        }}
-                      />
-                    )}
-                  </div>
-                  <br />
-                  <p
-                    className="text text--small"
-                    style={{
-                      display: 'inline-block',
-                      textAlign: 'justify',
-                      padding: '4px',
-                      border: '1px dashed',
-                      borderRadius: '4px',
-                      maxWidth: '480px',
-                    }}
-                  >
-                    Ne donnez pas le nom de votre {enigmeType.title.toLowerCase()}. La catégorie de {"l'énigme"} est un{' '}
-                    <strong>indice supplémentaire</strong> pour les autres classes.
-                  </p>
-                  <br />
-                  <Button color="primary" size="small" variant="outlined" onClick={onClick(-1)}>
-                    Continuer
-                  </Button>
-                </div>
-              }
-              label="Autre"
-              description={`Présentez ${enigmeType.title3} d’une autre catégorie.`}
-            />
+          <h1>Commencer un nouvel échange avec vos Pélicopains :</h1>
+          <div style={{ margin: '1rem 0 3rem 0' }}>
+            <ThemeChoiceButton label="Créer une nouvelle énigme" description="" onClick={onNext(true)} />
+          </div>
+          <h1>Réagir à une activité déjà publiée par vos Pélicopains :</h1>
+          <div ref={selectRef}>
+            <ActivitySelect value={activity.responseActivityId} onChange={onChange} onSelect={onNext(false)} style={{ margin: '1rem 0 0 0' }} />
           </div>
         </div>
       </div>
