+++ src/pages/creer-une-enigme/3.tsx
@@ -11,48 +11,34 @@ import { Steps } from 'src/components/Steps';
 import { ContentEditor } from 'src/components/activities/content';
 import { ActivityContext } from 'src/contexts/activityContext';
 import { capitalize } from 'src/utils';
+import { ActivityStatus } from 'types/activity.type';
 
 const EnigmeStep3: React.FC = () => {
   const router = useRouter();
   const { activity, updateActivity, addContent, deleteContent, save } = React.useContext(ActivityContext);
 
   const data = (activity?.data as EnigmeData) || null;
-  const indiceContentIndex = data?.indiceContentIndex ?? 0;
+  const isEdit = activity !== null && activity.id !== 0 && activity.status !== ActivityStatus.DRAFT;
+  const indiceContentIndex = Math.max(data?.indiceContentIndex ?? 0, 0);
 
-  const contentAdded = React.useRef(false);
   React.useEffect(() => {
     if (activity === null && !('activity-id' in router.query) && !sessionStorage.getItem('activity')) {
       router.push('/creer-une-enigme');
     } else if (activity && !isEnigme(activity)) {
       router.push('/creer-une-enigme');
     }
-
-    if (activity && isEnigme(activity)) {
-      if ((activity.data.indiceContentIndex ?? 0) > activity.processedContent.length) {
-        updateActivity({
-          data: {
-            ...activity.data,
-            indiceContentIndex: activity.processedContent.length,
-          },
-        });
-      }
-      if ((activity.data.indiceContentIndex ?? 0) === activity.processedContent.length && !contentAdded.current) {
-        contentAdded.current = true;
-        addContent('text');
-      }
-    }
-  }, [activity, router, updateActivity, addContent]);
+  }, [activity, router]);
 
   const updateContent = (content: EditorContent[]): void => {
-    updateActivity({ processedContent: [...activity.processedContent.slice(0, indiceContentIndex), ...content] });
+    updateActivity({ processedContent: [...content, ...activity.processedContent.slice(indiceContentIndex, activity.processedContent.length)] });
   };
-  const addIndiceContent = (type: EditorTypes, value?: string) => {
-    contentAdded.current = true;
-    addContent(type, value);
+  const addDescriptionContent = (type: EditorTypes, value?: string) => {
+    addContent(type, value, indiceContentIndex);
+    updateActivity({ data: { ...data, indiceContentIndex: indiceContentIndex + 1 } });
   };
-  const deleteIndiceContent = (index: number) => {
-    contentAdded.current = true; // delete means there were content already
-    deleteContent(indiceContentIndex + index);
+  const deleteDescriptionContent = (index: number) => {
+    deleteContent(index);
+    updateActivity({ data: { ...data, indiceContentIndex: indiceContentIndex - 1 } });
   };
 
   if (data === null || !isEnigme(activity)) {
@@ -66,25 +52,25 @@ const EnigmeStep3: React.FC = () => {
     <Base>
       <div style={{ width: '100%', padding: '0.5rem 1rem 1rem 1rem' }}>
         <Steps
-          steps={[
+          steps={(isEdit ? [] : ['Démarrer']).concat([
             data.theme === -1 ? capitalize(data.themeName ?? '') : enigmeData[data.theme]?.step ?? 'Choix de la catégorie',
             enigmeType.step2 ?? "Description de l'objet",
             "Création de l'indice",
             'Prévisualisation',
-          ]}
-          activeStep={2}
+          ])}
+          activeStep={isEdit ? 1 : 2}
         />
         <div className="width-900">
-          <h1>Créer votre indice</h1>
+          <h1>{enigmeType.titleStep2}</h1>
           <p className="text" style={{ fontSize: '1.1rem' }}>
-            Créez ici un <strong>indice</strong> pour faire deviner votre {enigmeType.titleStep2Short}. Vous pouvez ajouter du texte, une vidéo ou une
-            image à votre indice et vous pourrez le modifier à l’étape 4.
+            Decrivez ici votre {enigmeType.titleStep2Short}, il s’agira de la <strong>réponse</strong> partagée aux autres classes. Vous pouvez
+            ajouter du texte, une vidéo ou une image à votre description. Vous pourrez le modifier à l’étape 4.
           </p>
           <ContentEditor
-            content={activity.processedContent.slice(indiceContentIndex, activity.processedContent.length)}
+            content={activity.processedContent.slice(0, indiceContentIndex)}
             updateContent={updateContent}
-            addContent={addIndiceContent}
-            deleteContent={deleteIndiceContent}
+            addContent={addDescriptionContent}
+            deleteContent={deleteDescriptionContent}
             save={save}
           />
           <StepsButton prev="/creer-une-enigme/2" next="/creer-une-enigme/4" />
