+++ src/pages/creer-une-enigme/5.tsx
@@ -6,8 +6,9 @@ import Backdrop from '@material-ui/core/Backdrop';
 import Button from '@material-ui/core/Button';
 import CircularProgress from '@material-ui/core/CircularProgress';
 
-import { isPresentation } from 'src/activity-types/anyActivity';
-import { isThematique, PRESENTATION_THEMATIQUE } from 'src/activity-types/presentation.const';
+import { isEnigme } from 'src/activity-types/anyActivity';
+import { ENIGME_TYPES, ENIGME_DATA } from 'src/activity-types/enigme.const';
+import { EnigmeData } from 'src/activity-types/enigme.types';
 import { Base } from 'src/components/Base';
 import { StepsButton } from 'src/components/StepsButtons';
 import { Steps } from 'src/components/Steps';
@@ -16,6 +17,7 @@ import { ContentView } from 'src/components/activities/content/ContentView';
 import { EditButton } from 'src/components/buttons/EditButton';
 import { ActivityContext } from 'src/contexts/activityContext';
 import { useActivity } from 'src/services/useActivity';
+import { capitalize } from 'src/utils';
 import { ActivityStatus, ActivityType } from 'types/activity.type';
 
 const REACTIONS = {
@@ -26,20 +28,21 @@ const REACTIONS = {
   [ActivityType.QUESTION]: 'cette question',
 };
 
-const PresentationStep4: React.FC = () => {
+const EnigmeStep5: React.FC = () => {
   const router = useRouter();
   const { activity, save } = React.useContext(ActivityContext);
   const { activity: responseActivity } = useActivity(activity?.responseActivityId ?? -1);
   const [isLoading, setIsLoading] = React.useState(false);
 
-  const data = activity?.data || null;
+  const data = (activity?.data as EnigmeData) || null;
   const isEdit = activity !== null && activity.id !== 0 && activity.status !== ActivityStatus.DRAFT;
+  const indiceContentIndex = data?.indiceContentIndex ?? 0;
 
   React.useEffect(() => {
     if (activity === null && !('activity-id' in router.query) && !sessionStorage.getItem('activity')) {
-      router.push('/se-presenter');
-    } else if (activity && (!isPresentation(activity) || !isThematique(activity))) {
-      router.push('/se-presenter');
+      router.push('/creer-une-enigme');
+    } else if (activity && !isEnigme(activity)) {
+      router.push('/creer-une-enigme');
     }
   }, [activity, router]);
 
@@ -47,31 +50,42 @@ const PresentationStep4: React.FC = () => {
     setIsLoading(true);
     const success = await save(true);
     if (success) {
-      router.push('/se-presenter/success');
+      router.push('/creer-une-enigme/success');
     }
     setIsLoading(false);
   };
 
-  if (data === null || !('theme' in data) || data.theme === -1) {
+  if (data === null || !isEnigme(activity)) {
     return <div></div>;
   }
+
+  const enigmeType = ENIGME_TYPES[activity.subType ?? 0] ?? ENIGME_TYPES[0];
+  const enigmeData = ENIGME_DATA[activity.subType ?? 0] ?? ENIGME_DATA[0];
 
   return (
     <Base>
       <div style={{ width: '100%', padding: '0.5rem 1rem 1rem 1rem' }}>
-        <Steps steps={(isEdit ? [] : ['Démarrer']).concat(['Choix du thème', 'Présentation', 'Prévisualisation'])} activeStep={isEdit ? 2 : 3} />
+        <Steps
+          steps={(isEdit ? [] : ['Démarrer']).concat([
+            data.theme === -1 ? capitalize(data.themeName ?? '') : enigmeData[data.theme]?.step ?? 'Choix de la catégorie',
+            enigmeType.step2 ?? "Description de l'objet",
+            "Création de l'indice",
+            'Prévisualisation',
+          ])}
+          activeStep={isEdit ? 3 : 4}
+        />
         <div className="width-900">
-          <h1>Pré-visualisez votre présentation{!isEdit && ' et publiez la'}</h1>
+          <h1>Pré-visualisez votre énigme{!isEdit && ', et publiez-la'}</h1>
           <p className="text" style={{ fontSize: '1.1rem' }}>
-            Voici la pré-visualisation de votre présentation.
+            Voici la pré-visualisation de votre énigme.
             {isEdit
               ? " Vous pouvez la modifier à l'étape précédente, et enregistrer vos changements ici."
               : ' Vous pouvez la modifier, et quand vous êtes prêts : publiez-la dans votre village-monde !'}
           </p>
           {isEdit ? (
             <div style={{ width: '100%', display: 'flex', justifyContent: 'space-between', margin: '1rem 0' }}>
-              <Link href="/se-presenter/thematique/3">
-                <Button component="a" color="secondary" variant="contained" href="/se-presenter/thematique/3">
+              <Link href="/creer-une-enigme/4">
+                <Button component="a" color="secondary" variant="contained" href="/creer-une-enigme/4">
                   {"Modifier à l'étape précédente"}
                 </Button>
               </Link>
@@ -89,9 +103,9 @@ const PresentationStep4: React.FC = () => {
 
           {!isEdit && activity.responseActivityId === null && (
             <div style={{ margin: '1rem 0' }}>
-              Votre présentation initie un nouvel échange avec les Pélicopains,{' '}
-              <Link href={`/se-presenter/thematique/1?edit=${activity.id}`}>
-                <a className="text text--primary" href={`/se-presenter/thematique/1?edit=${activity.id}`}>
+              Votre énigme initie un nouvel échange avec les Pélicopains,{' '}
+              <Link href={`/creer-une-enigme/1?edit=${activity.id}`}>
+                <a className="text text--primary" href={`/creer-une-enigme/1?edit=${activity.id}`}>
                   si vous souhaitez plutôt réagir à une activité déjà publiée, cliquez ici.
                 </a>
               </Link>
@@ -99,12 +113,12 @@ const PresentationStep4: React.FC = () => {
           )}
           {responseActivity !== null && (
             <>
-              <span className="text text--small text--success">Présentation en réaction à {REACTIONS[responseActivity.type]}</span>
+              <span className="text text--small text--success">Énigme en réaction à {REACTIONS[responseActivity.type]}</span>
               <div className="preview-block">
                 {!isEdit && (
                   <EditButton
                     onClick={() => {
-                      router.push(`/se-presenter/thematique/1?edit=${activity.id}`);
+                      router.push(`/creer-une-enigme/1?edit=${activity.id}`);
                     }}
                     isGreen
                     style={{ position: 'absolute', top: '0.5rem', right: '0.5rem' }}
@@ -115,31 +129,46 @@ const PresentationStep4: React.FC = () => {
             </>
           )}
 
-          <span className="text text--small text--success">Thème</span>
+          <span className="text text--small text--success">{"Catégorie de l'énigme"}</span>
           <div className="preview-block">
             <EditButton
               onClick={() => {
-                router.push('/se-presenter/thematique/2');
+                router.push('/creer-une-enigme/2');
               }}
               isGreen
               style={{ position: 'absolute', top: '0.5rem', right: '0.5rem' }}
             />
-            <p style={{ margin: '0.5rem 0' }}>{PRESENTATION_THEMATIQUE[(data?.theme as number | null) ?? 0]?.cardTitle}</p>
+            <p style={{ margin: '0.5rem 0' }}>
+              Notre {enigmeType.titleStep2Short} mystère est{' '}
+              <strong>{(data.theme === -1 ? data.themeName ?? '' : enigmeData[data.theme]?.step ?? '').toLowerCase()}</strong>.
+            </p>
           </div>
 
-          <span className="text text--small text--success">Présentation</span>
+          <span className="text text--small text--success">{enigmeType.step2 ?? "Description de l'objet"}</span>
           <div className="preview-block">
             <EditButton
               onClick={() => {
-                router.push('/se-presenter/thematique/3');
+                router.push('/creer-une-enigme/3');
               }}
               isGreen
               style={{ position: 'absolute', top: '0.5rem', right: '0.5rem' }}
             />
-            <ContentView content={activity.processedContent} />
+            <ContentView content={activity.processedContent.slice(0, indiceContentIndex)} />
           </div>
 
-          <StepsButton prev="/se-presenter/thematique/3" />
+          <span className="text text--small text--success">Indice présenté aux autres classes</span>
+          <div className="preview-block">
+            <EditButton
+              onClick={() => {
+                router.push('/creer-une-enigme/4');
+              }}
+              isGreen
+              style={{ position: 'absolute', top: '0.5rem', right: '0.5rem' }}
+            />
+            <ContentView content={activity.processedContent.slice(indiceContentIndex, activity.processedContent.length)} />
+          </div>
+
+          <StepsButton prev="/creer-une-enigme/4" />
         </div>
       </div>
       <Backdrop style={{ zIndex: 2000, color: 'white' }} open={isLoading}>
@@ -149,4 +178,4 @@ const PresentationStep4: React.FC = () => {
   );
 };
 
-export default PresentationStep4;
+export default EnigmeStep5;
