+++ src/pages/se-presenter/thematique/2.tsx
@@ -9,12 +9,14 @@ import { StepsButton } from 'src/components/StepsButtons';
 import { Steps } from 'src/components/Steps';
 import { ThemeChoiceButton } from 'src/components/buttons/ThemeChoiceButton';
 import { ActivityContext } from 'src/contexts/activityContext';
+import { ActivityStatus } from 'types/activity.type';
 
 const PresentationStep2: React.FC = () => {
   const router = useRouter();
   const { activity, updateActivity } = React.useContext(ActivityContext);
 
   const data = (activity?.data as ThematiqueData) || null;
+  const isEdit = activity !== null && activity.id !== 0 && activity.status !== ActivityStatus.DRAFT;
 
   React.useEffect(() => {
     if (activity === null && !('activity-id' in router.query) && !sessionStorage.getItem('activity')) {
@@ -36,7 +38,7 @@ const PresentationStep2: React.FC = () => {
   return (
     <Base>
       <div style={{ width: '100%', padding: '0.5rem 1rem 1rem 1rem' }}>
-        <Steps steps={['Démarrer', 'Choix du thème', 'Présentation', 'Prévisualisation']} activeStep={1} />
+        <Steps steps={(isEdit ? [] : ['Démarrer']).concat(['Choix du thème', 'Présentation', 'Prévisualisation'])} activeStep={isEdit ? 0 : 1} />
         <div className="width-900">
           <h1>Choisissez le thème de votre présentation</h1>
           <p className="text" style={{ fontSize: '1.1rem' }}>
@@ -47,7 +49,7 @@ const PresentationStep2: React.FC = () => {
               <ThemeChoiceButton key={index} label={t.label} description={t.description} onClick={onClick(index)} />
             ))}
           </div>
-          <StepsButton prev={`/se-presenter/thematique/1?edit=${activity.id}`} />
+          {!isEdit && <StepsButton prev={`/se-presenter/thematique/1?edit=${activity.id}`} />}
         </div>
       </div>
     </Base>
